+++ Models/Devoluciones.cs
@@ -2,7 +2,7 @@
 {
     public class Devoluciones
     {
-        public decimal IdDevolucionPk { get; set; }
+        public int IdDevolucionPk { get; set; }
         public string? AlumnoDevolucion { get; set; }
         public int? Prestamo { get; set; }
         public DateTime? FechaDevolucion { get; set; }
